replaced reducer tests
